requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0

# 文档解析
langchain>=0.1.0
langchain-community>=0.0.20
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# 向量数据库和语义搜索
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0

# AI/ML工具
openai>=1.12.0  # 可选，用于AI功能
tiktoken>=0.5.0

# 前端集成
python-telegram-bot>=20.7
httpx>=0.25.0
aiohttp>=3.9.0
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0

# 管理界面
streamlit>=1.29.0
streamlit-option-menu>=0.3.6
plotly>=5.18.0
pandas>=2.1.0

# 工具库
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
